Refactor package strucure for messages
